Update naming, navigate to job offers screen #70
