+++ network.js
@@ -42,7 +42,7 @@ export async function signUp(
 //======================================To get user's posts ==========================================//
 export async function getAllPosts(uid) {
   try {
-    const res = await axios.get(`${url}/api/user/${uid}`);
+    const res = await axios.get(`${url}/api/posts/user/${uid}`);
     return res.data;
   } catch (err) {
     console.log(err);
