+++ src/components/ServiceProviderCard/ServiceProviderCard.js
@@ -1,0 +1,16 @@
+import React from 'react';
+import { StyleSheet, View, Text, Image, TouchableOpacity } from 'react-native';
+import { Card, Badge, Button } from 'react-native-elements';
+import { FlatList } from 'react-native-gesture-handler';
+
+export default function ServiceProviderCard({ }) {
+
+    return (
+        <Text>This is service provider card</Text>
+    )
+}
+
+const styles = StyleSheet.create({
+
+})
+
