+++ src/screens/index.js
@@ -2,5 +2,5 @@ export { default as ServiceScreen } from './ServiceScreen/ServiceScreen';
 
 export { default as NavigationScreen } from './NavigationScreen/NavigationScreen';
 
-export { default as Home } from './HomeScreen/Home';
+export { default as MyPostList } from './MyPostListScreen/MyPostList';
 export { default as AddItemScreen } from './AddItemScreen/AddItemScreen';
